Split up `exec/operator.hpp`
