+++ libtenzir/include/tenzir/exec/pipeline.hpp
@@ -8,24 +8,11 @@
 
 #pragma once
 
-#include "tenzir/exec/checkpoint_reader.hpp"
+#include "tenzir/exec/actors.hpp"
 #include "tenzir/exec/pipeline_settings.hpp"
 #include "tenzir/plan/pipeline.hpp"
 
 namespace tenzir::exec {
-
-struct pipeline_actor_traits {
-  using signatures = caf::type_list<
-    // Starts the pipeline, returning after the pipeline's startup sequence has
-    // completed. This handler requires the pipeline to be closed.
-    auto(atom::start)->caf::result<void>,
-    // Starts the pipeline with an existing handshake. The handshake's type must
-    // match the pipeline's input type. The handler returns the handshake from
-    // the pipeline's last operator.
-    auto(atom::start, handshake hs)->caf::result<handshake_response>>;
-};
-
-using pipeline_actor = caf::typed_actor<pipeline_actor_traits>;
 
 /// Create a new pipeline executor.
 ///
